requirements.txt
numpy
timm
black
jaxtyping
matplotlib
Pillow
scikit-learn
einops
rerun-sdk
pre-commit
minio
flow_vis
pytest
isort
